reafctor: merge LoginForm & RegistrationForm to AuthForm
