+++ src/pages/Authorization.tsx
@@ -1,17 +1,8 @@
 import React from "react";
-import { Box, FormControlLabel, Checkbox } from "@mui/material";
-import LoginForm from "components/LogInForm";
-import RegistrationForm from "components/RegistrationForm";
+import { Box } from "@mui/material";
+import AuthForm from "components/AuthForm";
 
 const Authorization: React.FC = () => {
-  const [isNew, setIsNew] = React.useState(false);
-
-  const handleChange = (event: {
-    target: { checked: boolean | ((prevState: boolean) => boolean) };
-  }) => {
-    setIsNew(event.target.checked);
-  };
-
   return (
     <Box
       sx={{ border: 1, borderColor: "grey.500" }}
@@ -21,11 +12,7 @@ const Authorization: React.FC = () => {
       height="auto"
       padding="20px"
     >
-      {isNew ? <RegistrationForm /> : <LoginForm />}
-      <FormControlLabel
-        control={<Checkbox checked={isNew} onChange={handleChange} />}
-        label="New user?"
-      />
+      <AuthForm />
     </Box>
   );
 };
